feat: имплементировать CRUD методы для UserStorage и FilmStorage feat: реализовать функции добавления и удаления друзей

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -34,7 +34,7 @@ public class FilmController {
     }
 
     @GetMapping("/{id}")
-    public Film findFilmBy(@PathVariable("id") Integer id) {
+    public Film findFilmBy(@PathVariable("id") Long id) {
         return filmService.getFilmBy(id);
     }
 }
